move widgets in seperate folders
