+++ lib/screens/interests/interest_pages_screen.dart
@@ -11,7 +11,7 @@ import 'package:gruene_app/common/utils/image_provider_delegate.dart';
 import 'package:gruene_app/net/interests/bloc/interests_bloc.dart';
 import 'package:gruene_app/net/interests/repository/interests_repository.dart';
 import 'package:gruene_app/routing/routes.dart';
-import 'package:gruene_app/widget/page_stepper.dart';
+import 'package:gruene_app/widget/steppers/page_stepper.dart';
 
 class InterestPagesScreen extends StatelessWidget {
   const InterestPagesScreen({super.key});
