+++ lib/screens/interests/interest_start_screen.dart
@@ -6,7 +6,7 @@ import 'package:gruene_app/gen/assets.gen.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:gruene_app/routing/app_startup.dart';
 import 'package:gruene_app/routing/routes.dart';
-import 'package:gruene_app/widget/button_group.dart';
+import 'package:gruene_app/widget/buttons/button_group.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:vector_graphics/vector_graphics.dart';
