+++ lib/screens/more/more_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gruene_app/constants/theme_data.dart';
 import 'package:gruene_app/routing/routes.dart';
-import 'package:gruene_app/widget/costume_separated_list.dart';
+import 'package:gruene_app/widget/lists/costume_separated_list.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class MoreScreen extends StatelessWidget {
