+++ lib/screens/more/screens/profile/member_profil_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gruene_app/constants/theme_data.dart';
 import 'package:gruene_app/net/profile/bloc/profile_bloc.dart';
 import 'package:gruene_app/net/profile/data/member_profil.dart';
-import 'package:gruene_app/widget/costume_separated_list.dart';
+import 'package:gruene_app/widget/lists/costume_separated_list.dart';
 import 'package:gruene_app/widget/modals/multi_modal_select.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
