+++ lib/screens/more/screens/profile/profile_menu.dart
@@ -7,7 +7,7 @@ import 'package:gruene_app/net/authentication/authentication.dart';
 import 'package:gruene_app/net/profile/bloc/profile_bloc.dart';
 import 'package:gruene_app/routing/router.dart';
 import 'package:gruene_app/routing/routes.dart';
-import 'package:gruene_app/widget/costume_separated_list.dart';
+import 'package:gruene_app/widget/lists/costume_separated_list.dart';
 import 'package:gruene_app/screens/more/screens/profile/profile_list_view_header.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
